requirements.txt
asgiref==3.8.1
boto==2.49.0
Django==5.1.7
django-crispy-forms==2.3
django-dynamodb-sessions==0.6
django-storages==1.14.5
sqlparse==0.5.3
typing_extensions==4.13.0
